Remove db from Store
